requirements.txt
pandas
numpy
matplotlib
scikit-learn
sentence-transformers
faiss-cpu
fastapi
uvicorn
streamlit
nltk
joblib
pyidaungsu
